app/api/requirements.txt
-i https://pypi.org/simple
anyio==3.6.1; python_full_version >= '3.6.2'
beautifulsoup4==4.11.1; python_version >= '3.6'
bs4==0.0.1
certifi==2022.6.15; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
elastic-transport==8.1.2; python_version >= '3.6'
elasticsearch==8.3.3
fastapi==0.79.0
h11==0.13.0; python_version >= '3.6'
idna==3.3; python_version >= '3.5'
pydantic==1.9.2; python_full_version >= '3.6.1'
python-dotenv==0.20.0
sniffio==1.2.0; python_version >= '3.5'
soupsieve==2.3.2.post1; python_version >= '3.6'
sourcery==1.2.7
starlette==0.19.1; python_version >= '3.6'
typing-extensions==4.3.0; python_version >= '3.7'
urllib3==1.26.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
uvicorn==0.18.2
